✨ feat: Implement TimeDealController and update UserController
- 타임딜 등록 API 추가 ()
- UserController에 타임딜 관리자 검증 로직 추가

+++ src/main/java/goorm/server/timedeal/controller/UserController.java
@@ -24,7 +24,7 @@ public class UserController {
 
 	// User 수정 API
 	@PutMapping("/{id}")
-	public ResponseEntity<User> updateUser(@PathVariable int id, @Validated @RequestBody User updatedUser) {
+	public ResponseEntity<User> updateUser(@PathVariable Long id, @Validated @RequestBody User updatedUser) {
 		User user = userService.updateUser(id, updatedUser);
 		return ResponseEntity.ok(user);
 	}
